requirements.txt
python-dotenv 
pandas 
numpy
pymongo
certifi
pymongo>=4.0
scikit-learn
dill
pyaml
mlflow
dagshub
fastapi
uvicorn
python-multipart
ipykernel
-e .